defined register post request
